+++ app/src/js/components/Executions/index.js
@@ -1,7 +1,11 @@
 'use strict';
 import React from 'react';
+import { withRouter, Route, Switch } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import Sidebar from '../Sidebar/sidebar';
+import ExecutionOverview from './overview';
+import ExecutionStatus from './execution-status';
+import ExecutionLogs from './execution-logs';
 
 class Executions extends React.Component {
   render () {
@@ -19,7 +23,11 @@ class Executions extends React.Component {
               params={this.props.params}
             />
             <div className='page__content--shortened'>
-              {this.props.children}
+              <Switch>
+                <Route exact path='/executions' component={ExecutionOverview} />
+                <Route path='/executions/execution/:executionName/logs' component={ExecutionLogs} />
+                <Route path='/executions/execution/:executionArn' component={ExecutionStatus} />
+              </Switch>
             </div>
           </div>
         </div>
@@ -34,4 +42,4 @@ Executions.propTypes = {
   params: PropTypes.object
 };
 
-export default Executions;
+export default withRouter(Executions);