+++ app/src/js/components/Providers/edit.js
@@ -2,6 +2,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
+import { withRouter } from 'react-router-dom';
 import {
   getProvider,
   updateProvider,
@@ -13,7 +14,7 @@ const SCHEMA_KEY = 'provider';
 
 class EditProvider extends React.Component {
   render () {
-    const { providerId } = this.props.params;
+    const { providerId } = this.props.match.params;
     return (
       <EditRecord
         merge={true}
@@ -23,17 +24,17 @@ class EditProvider extends React.Component {
         getRecord={getProvider}
         updateRecord={updateProvider}
         clearRecordUpdate={clearUpdateProvider}
-        backRoute={`providers/provider/${providerId}`}
+        backRoute={`/providers/provider/${providerId}`}
       />
     );
   }
 }
 
 EditProvider.propTypes = {
-  params: PropTypes.object,
+  match: PropTypes.object,
   providers: PropTypes.object
 };
 
-export default connect(state => ({
+export default withRouter(connect(state => ({
   providers: state.providers
-}))(EditProvider);
+}))(EditProvider));