+++ app/src/js/components/Rules/index.js
@@ -1,8 +1,12 @@
 'use strict';
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Link } from 'react-router';
+import { withRouter, Link, Route, Switch } from 'react-router-dom';
 import Sidebar from '../Sidebar/sidebar';
+import RulesOverview from './overview';
+import Rule from './rule';
+import EditRule from './edit';
+import AddRule from './add';
 
 class Rules extends React.Component {
   render () {
@@ -13,19 +17,24 @@ class Rules extends React.Component {
         <div className='content__header'>
           <div className='row'>
             <h1 className='heading--xlarge heading--shared-content'>Rules</h1>
-            { showSidebar ? <Link className='button button--large button--white button__addcollections button__arrow button__animation' to='/rules/add'>Add a rule</Link> : null }
+            {showSidebar ? <Link className='button button--large button--white button__addcollections button__arrow button__animation' to='/rules/add'>Add a rule</Link> : null}
           </div>
         </div>
         <div className='page__content'>
           <div className='wrapper__sidebar'>
-            { showSidebar ? (
+            {showSidebar ? (
               <Sidebar
                 currentPath={this.props.location.pathname}
                 params={this.props.params}
               />
-            ) : null }
-            <div className={ showSidebar ? 'page__content--shortened' : 'page__content' }>
-              {this.props.children}
+            ) : null}
+            <div className={showSidebar ? 'page__content--shortened' : 'page__content'}>
+              <Switch>
+                <Route exact path='/rules' component={RulesOverview} />
+                <Route path='/rules/rule/:ruleName' component={Rule} />
+                <Route path='/rules/edit/:ruleName' component={EditRule} />
+                <Route path='/rules/add' component={AddRule} />
+              </Switch>
             </div>
           </div>
         </div>
@@ -40,4 +49,4 @@ Rules.propTypes = {
   params: PropTypes.object
 };
 
-export default Rules;
+export default withRouter(Rules);