+++ app/src/js/components/Search/search.js
@@ -1,9 +1,10 @@
 'use strict';
 import React from 'react';
-import { withRouter } from 'react-router';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { initialValueFromLocation, updateRouterLocation } from '../../utils/url-helper';
+import { initialValueFromLocation } from '../../utils/url-helper';
+import withQueryParams from 'react-router-query-params';
+import { withRouter } from 'react-router-dom';
 
 class Search extends React.Component {
   constructor (props) {
@@ -17,13 +18,18 @@ class Search extends React.Component {
   }
 
   componentDidMount () {
-    const { dispatch, action, location, paramKey } = this.props;
-    dispatch(action({value: location.query[paramKey]}));
+    const { dispatch, action, paramKey, queryParams } = this.props;
+    const queryValue = queryParams[paramKey];
+    if (queryValue) dispatch(action({value: queryValue}));
   }
 
-  componentDidUpdate () {
-    const { dispatch, action, location, paramKey } = this.props;
-    dispatch(action({value: location.query[paramKey]}));
+  componentDidUpdate (prevProps, prevState, snapshot) {
+    const {paramKey, dispatch, action, queryParams} = this.props;
+    const queryValue = queryParams[paramKey];
+    if (queryValue !== prevProps.queryParams[paramKey]) {
+      dispatch(action({ key: paramKey, queryValue }));
+      this.setState({queryValue}); // eslint-disable-line react/no-did-update-set-state
+    }
   }
 
   componentWillUnmount () {
@@ -49,13 +55,13 @@ class Search extends React.Component {
   }
 
   delayedQuery (value) {
-    const { dispatch, action, clear, location, router, paramKey } = this.props;
+    const { dispatch, action, clear, paramKey, setQueryParams } = this.props;
     const timeoutId = setTimeout(function () {
       if (value && value.length) {
-        updateRouterLocation(router, location, paramKey, value);
+        setQueryParams({[paramKey]: value});
         dispatch(action(value));
       } else {
-        updateRouterLocation(router, location, paramKey, '');
+        setQueryParams({[paramKey]: ''});
         dispatch(clear());
       }
     }, 650);
@@ -66,13 +72,13 @@ class Search extends React.Component {
     const { label, paramKey } = this.props;
     const formID = `form-${label}-${paramKey}`;
     return (
-     <div className='filter__item'>
-       {label ? <label htmlFor={formID}>{label}</label> : null}
-       <form className='search__wrapper form-group__element' onSubmit={this.submit} >
-        <input className='search' type='search' onChange={this.complete} value={this.state.value}/>
-        <span className='search__icon'/>
-       </form>
-     </div>
+      <div className='filter__item'>
+        {label ? <label htmlFor={formID}>{label}</label> : null}
+        <form className='search__wrapper form-group__element' onSubmit={this.submit} >
+          <input className='search' type='search' onChange={this.complete} value={this.state.value || ''}/>
+          <span className='search__icon'/>
+        </form>
+      </div>
     );
   }
 }
@@ -88,7 +94,10 @@ Search.propTypes = {
   paramKey: PropTypes.string,
   label: PropTypes.any,
   location: PropTypes.object,
-  router: PropTypes.object
+  router: PropTypes.object,
+  query: PropTypes.object,
+  queryParams: PropTypes.object,
+  setQueryParams: PropTypes.func
 };
 
-export default withRouter(connect()(Search));
+export default withRouter(withQueryParams()(connect(state => state)(Search)));