+++ app/src/js/components/home.js
@@ -4,7 +4,7 @@ import moment from 'moment';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { Link, withRouter } from 'react-router';
+import { withRouter, Link } from 'react-router-dom';
 import { get } from 'object-path';
 import {
   getCount,
@@ -31,7 +31,7 @@ import {
   errorTableRow,
   errorTableSortProps
 } from '../utils/table-config/granules';
-import { recent, updateInterval } from '../config';
+import _config from '../config';
 import {
   kibanaS3AccessErrorsLink,
   kibanaS3AccessSuccessesLink,
@@ -48,24 +48,27 @@ import { initialValuesFromLocation } from '../utils/url-helper';
 import Datepicker from './Datepicker/Datepicker';
 import { strings } from './locale';
 
-/**
- * If this is a shared URL. grab the date and time and update the datepicker
- * state to reflect the values.
- * @param {Object} props - Home component's input props.
- */
-const updateDatepickerStateFromUrl = (props) => {
-  const urlDateFormat = 'YYYYMMDDHHmmSS';
-  const urlProps = ['endDateTime', 'startDateTime'];
-  const { location } = props;
-  const values = initialValuesFromLocation(location, urlProps);
-  if (!isEmpty(values)) {
-    for (const value in values) {
-      values[value] = moment.utc(values[value], urlDateFormat).toDate();
-    }
-    values.dateRange = {value: 'Custom', label: 'Custom'};
-    props.dispatch({type: 'DATEPICKER_DATECHANGE', data: {...props.datepicker, ...values}});
-  }
-};
+// TODO [MHS, 2020-02-19]  Hows this going to work?
+// /**
+//  * If this is a shared URL. grab the date and time and update the datepicker
+//  * state to reflect the values.
+//  * @param {Object} props - Home component's input props.
+//  */
+// const updateDatepickerStateFromUrl = (props) => {
+//   const urlDateFormat = 'YYYYMMDDHHmmSS';
+//   const urlProps = ['endDateTime', 'startDateTime'];
+//   const { location } = props;
+//   const values = initialValuesFromLocation(location, urlProps);
+//   if (!isEmpty(values)) {
+//     for (const value in values) {
+//       values[value] = moment.utc(values[value], urlDateFormat).toDate();
+//     }
+//     values.dateRange = {value: 'Custom', label: 'Custom'};
+//     props.dispatch({type: 'DATEPICKER_DATECHANGE', data: {...props.datepicker, ...values}});
+//   }
+// };
+
+const { recent, updateInterval } = _config;
 
 class Home extends React.Component {
   constructor (props) {
@@ -79,8 +82,9 @@ class Home extends React.Component {
     this.cancelInterval = interval(() => {
       this.query();
     }, updateInterval, true);
-    const {dispatch} = this.props;
-    updateDatepickerStateFromUrl(this.props);
+    // TODO [MHS, 2020-02-19] Fix this too
+    // updateDatepickerStateFromUrl(this.props);
+    const { dispatch } = this.props;
     dispatch(getCumulusInstanceMetadata())
       .then(() => {
         dispatch(getDistApiGatewayMetrics(this.props.cumulusInstance));
@@ -129,26 +133,28 @@ class Home extends React.Component {
       <section className='page__section'>
         <div className='row'>
           <div className='heading__wrapper'>
-              <h2 className='heading--medium heading--shared-content--right'>{header}</h2>
+            <h2 className='heading--medium heading--shared-content--right'>{header}</h2>
           </div>
-          <ul id={listId}>
-            {data.map(d => {
-              const value = d[0];
-              return (
+          <div className="overview-num__wrapper-home">
+            <ul id={listId}>
+              {data.map(d => {
+                const value = d[0];
+                return (
                   <li key={d[1]}>
-                  {this.isExternalLink(d[2]) ? (
-                    <a id={d[1]} href={d[2]} className='overview-num' target='_blank'>
-                      <span className='num--large'>{value}</span> {d[1]}
-                    </a>
-                  ) : (
-                    <Link id={d[1]} className='overview-num' to={d[2] || '#'}>
-                      <span className='num--large'>{value}</span> {d[1]}
-                    </Link>
-                  )}
-                </li>
-              );
-            })}
-          </ul>
+                    {this.isExternalLink(d[2]) ? (
+                      <a id={d[1]} href={d[2]} className='overview-num' target='_blank'>
+                        <span className='num--large'>{value}</span> {d[1]}
+                      </a>
+                    ) : (
+                      <Link id={d[1]} className='overview-num' to={d[2] || '#'}>
+                        <span className='num--large'>{value}</span> {d[1]}
+                      </Link>
+                    )}
+                  </li>
+                );
+              })}
+            </ul>
+          </div>
         </div>
       </section>
     );
@@ -271,6 +277,7 @@ Home.propTypes = {
 };
 
 export { Home };
+
 export default withRouter(connect((state) => ({
   cumulusInstance: state.cumulusInstance,
   datepicker: state.datepicker,