+++ app/src/js/reducers/index.js
@@ -1,4 +1,5 @@
 import { combineReducers } from 'redux';
+import { connectRouter } from 'connected-react-router';
 import api from './api';
 import apiVersion from './api-version';
 import collections from './collections';
@@ -50,4 +51,10 @@ export const reducers = {
   reconciliationReports
 };
 
-export default combineReducers(reducers);
+export const createRootReducer = (history) => combineReducers({
+  router: connectRouter(history),
+  ...reducers
+});
+
+//TODO [MHS, 2020-02-19]  Can this be deleted now?  Are we still importing js/reducers?
+export default combineReducers(Object.assign({}, reducers));