+++ cypress/plugins/index.js
@@ -11,7 +11,7 @@
 // This function is called when a project is opened or re-opened (e.g. due to
 // the project's config changing)
 
-const browserify = require('@cypress/browserify-preprocessor');
+const webpack = require('@cypress/webpack-preprocessor');
 
 const { testUtils } = require('@cumulus/api');
 
@@ -20,19 +20,17 @@ const { seedEverything } = require('./seedEverything');
 const fakeApiToken = require('./token');
 
 module.exports = (on) => {
-  const options = browserify.defaultOptions;
-  const babelOptions = options.browserifyOptions.transform[1][1];
+  const options = {
+    // send in the options from your webpack.config.js, so it works the same
+    // as your app's code
+    webpackOptions: require('../../webpack.common'),
+    watchOptions: {},
+  };
   const user = 'testUser';
-  let esClient;
-  let esIndex;
-  babelOptions.global = true;
-  // ignore all node_modules except files in @cumulus/
-  // see https://github.com/cypress-io/cypress-browserify-preprocessor/issues/19
-  babelOptions.ignore = [/\/node_modules\/(?!@cumulus\/)/];
 
   // Run specialized file preprocessor to transpile ES6+ -> ES5
   // This fixes compatibility issues with Electron
-  on('file:preprocessor', browserify(options));
+  on('file:preprocessor', webpack(options));
 
   on('task', {
     resetState: function () {